requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.16.0
newspaper3k>=0.2.8

# Data processing
pandas>=2.2.0
numpy>=1.26.2
nltk>=3.8.1
scikit-learn>=1.3.2
spacy>=3.7.2
regex>=2023.10.3

# Deep learning - Apple Silicon specific
tensorflow-macos>=2.15.0
tensorflow-metal>=1.0.0  # GPU acceleration for Mac
transformers>=4.38.2
torch>=2.2.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
schedule>=1.2.1